Set a sane HOME for binarypkg jobs
The comment above this invocation indicates that apt needs HOME to be
set to something, but the variable is set to an empty string. This is
even more critical for running the buildfarm using rootless Podman,
because `/` is no longer writable and seems to be the default of HOME is
empty.